+++ packages/go_router/example/lib/stateful_shell_route.dart
@@ -88,7 +88,7 @@ class NestedTabNavigationExampleApp extends StatelessWidget {
                     builder: (BuildContext context, GoRouterState state) =>
                         DetailsScreen(
                       label: 'B',
-                      param: state.params['param'],
+                      param: state.pathParameters['param'],
                       extra: state.extra,
                     ),
                   ),