+++ packages/go_router/lib/src/delegate.dart
@@ -27,7 +27,7 @@ class GoRouterDelegate extends RouterDelegate<RouteMatchList>
     required List<NavigatorObserver> observers,
     required this.routerNeglect,
     String? restorationScopeId,
-  })  : _configuration = configuration {
+  }) : _configuration = configuration {
     builder = RouteBuilder(
       configuration: configuration,
       builderWithNav: builderWithNav,
@@ -92,8 +92,12 @@ class GoRouterDelegate extends RouterDelegate<RouteMatchList>
       matches: matches,
     );
 
-    _matchList.push(newPageKeyMatch);
-    return newPageKeyMatch._future;
+    _matchList = _matchList.push(newPageKeyMatch);
+    return newPageKeyMatch.future;
+  }
+
+  void _remove(RouteMatch match) {
+    _matchList = _matchList.remove(match);
   }
 
   /// Pushes the given location onto the page stack.
@@ -107,7 +111,7 @@ class GoRouterDelegate extends RouterDelegate<RouteMatchList>
   Future<T?> push<T extends Object?>(RouteMatchList matches) async {
     assert(matches.last.route is! ShellRoute);
 
-    final ValueKey<String> pageKey = _getNewKeyForPath(matches.fullpath);
+    final ValueKey<String> pageKey = _getNewKeyForPath(matches.fullPath);
     final Future<T?> future = _push(matches, pageKey);
     notifyListeners();
     return future;
@@ -153,7 +157,7 @@ class GoRouterDelegate extends RouterDelegate<RouteMatchList>
     if (match is ImperativeRouteMatch) {
       match.complete(result);
     }
-    _matchList.remove(match!);
+    _remove(match!);
     notifyListeners();
     assert(() {
       _debugAssertMatchListNotEmpty();
@@ -173,7 +177,7 @@ class GoRouterDelegate extends RouterDelegate<RouteMatchList>
   ///   state and not run any page animation.
   void pushReplacement(RouteMatchList matches) {
     assert(matches.last.route is! ShellRoute);
-    _matchList.remove(_matchList.last);
+    _remove(_matchList.last);
     push(matches); // [push] will notify the listeners.
   }
 
@@ -191,7 +195,7 @@ class GoRouterDelegate extends RouterDelegate<RouteMatchList>
     assert(matches.last.route is! ShellRoute);
     final RouteMatch routeMatch = _matchList.last;
     final ValueKey<String> pageKey = routeMatch.pageKey;
-    _matchList.remove(routeMatch);
+    _remove(routeMatch);
     _push(matches, pageKey);
     notifyListeners();
   }
@@ -309,35 +313,4 @@ class _NavigatorStateIterator extends Iterator<NavigatorState> {
     current = root;
     return true;
   }
-}
-
-/// The route match that represent route pushed through [GoRouter.push].
-class ImperativeRouteMatch<T> extends RouteMatch {
-  /// Constructor for [ImperativeRouteMatch].
-  ImperativeRouteMatch({
-    required super.pageKey,
-    required this.matches,
-  })  : _completer = Completer<T?>(),
-        super(
-          route: matches.last.route,
-          subloc: matches.last.subloc,
-          extra: matches.last.extra,
-          error: matches.last.error,
-        );
-
-  /// The matches that produces this route match.
-  final RouteMatchList matches;
-
-  /// The completer for the future returned by [GoRouter.push].
-  final Completer<T?> _completer;
-
-  /// Called when the corresponding [Route] associated with this route match is
-  /// completed.
-  void complete([dynamic value]) {
-    _completer.complete(value as T?);
-  }
-
-  /// The future of the [RouteMatch] completer.
-  /// When the future completes, this will return the value passed to [complete].
-  Future<T?> get _future => _completer.future;
 }